fixed dso citizen
