+++ packages/modules/fsm/src/components/inbox/FSMLink.js
@@ -17,7 +17,7 @@ const FSMLink = ({ isMobile, data }) => {
       notAccessTo: [COLLECTOR],
     },
     // { text: t("ES_TITLE_REPORTS"), link: "/employee" },
-    { text: t("ES_TITLE_DASHBOARD"), link: DSO ? "/digit-ui/employee/fsm/dso-dashboard" : "/employee", hyperlink: !DSO },
+    { text: t("ES_TITLE_DASHBOARD"), link: DSO ? "/digit-ui/citizen/fsm/dso-dashboard" : "/employee", hyperlink: !DSO },
   ];
 
   const [links, setLinks] = useState([]);
